subscription shared prepare for landing page
