+++ src/components/subscription/subscription-slide.tsx
@@ -1,190 +1,26 @@
 import React from "react";
 import styled, { ThemeProvider } from "styled-components";
 
-import { SUBSCRIPTION_DESKTOP_WIDTH, MOBILE_WIDTH, theme } from "../../config";
+import { SUBSCRIPTION_DESKTOP_WIDTH, theme } from "../../config";
 
 import ScreenContainer from "../shared/screen-container";
+import Subscription from "../shared/subscription";
 
-const FlexGrid = styled.div`
+const Container = styled.div`
   display: flex;
-  justify-content: space-between;
-  margin: 0 0 20px 0;
+  justify-content: center;
+  margin: 15px 0 20px 0;
   @media (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) {
     display: block;
-  }
-`;
-
-const FlexCol = styled.div`
-  width: 33%;
-  height: 451px;
-  background-color: ${props => props.theme.container.background};
-  padding-top: 15px;
-  margin-inline-end: 10px;
-  @media (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) {
-    width: 100%;
-    border-bottom: 1px solid #8faacc;
-    &:last-child {
-      border-bottom: none;
-    }
-  }
-  @media only screen and (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) and (min-width: ${MOBILE_WIDTH}px) {
-    height: 230px;
-  }
-`;
-
-const Title = styled.h1`
-  font-size: ${props => props.theme.container.title.size}px;
-  font-weight: bold;
-  font-style: ${props => props.theme.fontStyle};
-  font-stretch: ${props => props.theme.fontStretch};
-  line-height: ${props => props.theme.lineHeight};
-  letter-spacing: ${props => props.theme.letterSpacing};
-  color: ${props => props.theme.title.color};
-  margin: auto;
-  text-align: center;
-  margin-top: 35px;
-  @media only screen and (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) and (min-width: ${MOBILE_WIDTH}px) {
-    text-align: left;
-    margin: 20px 20px 0 20px;
-    width: 250px;
-    flex: 0.5;
-  }
-`;
-
-const Hr = styled.div`
-  width: ${props => props.theme.container.title.underline.width}px;
-  border-top: ${props => props.theme.container.title.underline.height}px solid
-    ${props => props.theme.container.title.underline.color};
-  margin: auto;
-  margin-top: 5px;
-  margin-bottom: 15px;
-  @media only screen and (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) and (min-width: ${MOBILE_WIDTH}px) {
-    margin: 5px 20px 0 20px;
-  }
-`;
-
-const Content = styled.p`
-  width: 171px;
-  min-height: 70px;
-  font-size: 12px;
-  font-weight: ${props => props.theme.fontWeight};
-  font-style: ${props => props.theme.fontStyle};
-  font-stretch: ${props => props.theme.fontStretch};
-  line-height: ${props => props.theme.lineHeight};
-  letter-spacing: ${props => props.theme.letterSpacing};
-  color: ${props => props.theme.container.content};
-  margin: 15px 15px 0 35px;
-  @media (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) {
-    width: auto;
-    margin: 0 30px 0 30px;
-  }
-  @media only screen and (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) and (min-width: ${MOBILE_WIDTH}px) {
-    margin: 15px 30px 0 20px;
-    width: 250px;
-  }
-`;
-
-const ContentBold = styled(Content)`
-  font-weight: bold;
-  min-height: 28px;
-  @media only screen and (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) and (min-width: ${MOBILE_WIDTH}px) {
-    display: none;
-  }
-`;
-
-const Price = styled.p`
-  width: 90px;
-  min-height: 50px;
-  font-size: ${props => props.theme.container.title.size}px;
-  font-weight: bold;
-  font-style: ${props => props.theme.fontStyle};
-  font-stretch: ${props => props.theme.fontStretch};
-  line-height: ${props => props.theme.lineHeight};
-  letter-spacing: ${props => props.theme.letterSpacing};
-  color: ${props => props.theme.container.content};
-  text-align: center;
-  margin: auto;
-  margin-top: 20px;
-  @media only screen and (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) and (min-width: ${MOBILE_WIDTH}px) {
-    position: relative;
-    top: -92px;
-    right: -100px;
-  }
-`;
-
-const Button = styled.button`
-  width: 171px;
-  height: 40px;
-  background-color: ${props => props.theme.button.background};
-  font-size: 16px;
-  font-weight: bold;
-  font-style: ${props => props.theme.fontStyle};
-  font-stretch: ${props => props.theme.fontStretch};
-  line-height: ${props => props.theme.lineHeight};
-  letter-spacing: ${props => props.theme.letterSpacing};
-  color: ${props => props.theme.button.color};
-  text-align: center;
-  margin: auto;
-  border: none;
-`;
-
-const ButtonWrapper = styled.div`
-  text-align: center;
-  margin: 20px 0 40px 0;
-  @media only screen and (max-width: ${SUBSCRIPTION_DESKTOP_WIDTH}px) and (min-width: ${MOBILE_WIDTH}px) {
-    position: relative;
-    top: -92px;
-    right: -150px;
   }
 `;
 
 const SubscriptionSlide = () => (
   <ThemeProvider theme={theme}>
     <ScreenContainer title={"Choose Subscription Plan"}>
-      <FlexGrid>
-        <FlexCol>
-          <Title>Basic</Title>
-          <Hr />
-          <Content>
-            Secure, encrypted storage solution perfect for the needs of the
-            individual.
-          </Content>
-          <ContentBold>128 GB secure, decentralized storage</ContentBold>
-          <ContentBold>Unlimited downloads</ContentBold>
-          <Price>2 OPQ</Price>
-          <ButtonWrapper>
-            <Button>CHOOSE PLAN</Button>
-          </ButtonWrapper>
-        </FlexCol>
-        <FlexCol>
-          <Title>Professional</Title>
-          <Hr />
-          <Content>
-            For professionals looking for a secure, easily accessible storage
-            solution while on the move.
-          </Content>
-          <ContentBold>512 GB secure, decentralized storage</ContentBold>
-          <ContentBold>Unlimited downloads</ContentBold>
-          <Price>8 OPQ</Price>
-          <ButtonWrapper>
-            <Button>CHOOSE PLAN</Button>
-          </ButtonWrapper>
-        </FlexCol>
-        <FlexCol>
-          <Title>Business</Title>
-          <Hr />
-          <Content>
-            A secure, encrypted storage solution for growing businesses. Perfect
-            for small teams.
-          </Content>
-          <ContentBold>1 TB secure, decentralized storage</ContentBold>
-          <ContentBold>Unlimited downloads</ContentBold>
-          <Price>16 OPQ</Price>
-          <ButtonWrapper>
-            <Button>CHOOSE PLAN</Button>
-          </ButtonWrapper>
-        </FlexCol>
-      </FlexGrid>
+      <Container>
+        <Subscription />
+      </Container>
     </ScreenContainer>
   </ThemeProvider>
 );
